// vysnhub-app/lib/types/product.ts
export interface VysnProduct {
  vysnName: string;
  itemNumberVysn: string;
  shortDescription: string;
  longDescription: string;
  weightKg: number;
  packagingWeightKg: number;
  grossWeightKg: number;
  manuallink?: string;
  barcodeNumber: number;
  installationDiameter?: string;
  cableLengthMm?: number;
  diameterMm?: number;
  lengthMm?: number;
  widthMm?: number;
  heightMm?: number;
  housingColor: string;
  grossPrice: number;
  katalogQ424?: number;
  category1: string;
  category2: string;
  groupName: string;
  lightDirection?: string;
  lumen?: number;
  driverInfo?: string;
  beamAngle?: number;
  packagingWidthMm?: number;
  packagingLengthMm?: number;
  beamAngleRange?: string;
  packagingHeightMm?: number;
  lightsource?: string;
  luminosityDecrease?: string;
  steering?: string;
  ledChipLifetime?: number;
  energyClass?: string;
  ingressProtection?: string;
  cct?: number;
  hscode?: number;
  socketInformationForRetrofitProducts?: string;
  cri?: number;
  wattage?: number;
  ledtype?: string;
  sdcm?: string;
  operatingMode?: string;
  lumenPerWatt?: number;
  numberOfSockets?: number;
  ugr?: string;
  cctswitchValue?: string;
  protectionClass?: string;
  baseSocket?: string;
  installation?: string;
  eprelpicturelink?: string;
  replaceableLightSource?: string;
  material?: string;
  impactResistance?: string;
  eprelLink?: string;
  coverable?: string;
  packagingUnitsDefaultIs1?: number;
  powerswitchValue?: string;
  countryOfOrigin?: string;
  product_picture_1?: string;
  product_picture_2?: string;
  product_picture_3?: string;
  product_picture_4?: string;
  product_picture_5?: string;
  product_picture_6?: string;
  product_picture_7?: string;
  product_picture_8?: string;
}